Modify solution to sort in feasible time
A modified merge sort is beeing implemented, in order to optimize
sorting time.

This mergesort is a hybrid between the 'classic' merge sort algorithm
and the insertion sort algorithm. If the list size is lesser than 5,
the solution will apply a insertion sort on the list, instead of a
merge sort (when the size would get to 1).

Both merge sort and insertion sort implementations here are sorting
in descending order. The related changes are highlighted as comments
in the solution.